Move `Lifetime` from `types` to `microgrid`
This is to match where `Lifetime` is defined in the protobuf messages.
