+++ src/frequenz/client/common/microgrid/_lifetime.py
