+++ src/frequenz/client/common/microgrid/proto/v1alpha8/__init__.py
@@ -3,8 +3,10 @@
 
 """Conversion of microgrid objects from/to protobuf v1alpha8."""
 
+from ._lifetime import lifetime_from_proto
 from ._microgrid import microgrid_from_proto
 
 __all__ = [
+    "lifetime_from_proto",
     "microgrid_from_proto",
 ]
