+++ src/frequenz/client/common/microgrid/proto/v1alpha8/_lifetime.py
