+++ src/frequenz/client/common/types/proto/v1alpha8/__init__.py
@@ -3,10 +3,8 @@
 
 """Common type protobuf v1alpha8 conversions."""
 
-from ._lifetime import lifetime_from_proto
 from ._location import location_from_proto
 
 __all__ = [
-    "lifetime_from_proto",
     "location_from_proto",
 ]
